+++ docs/_static/searchtools.js
@@ -10,6 +10,32 @@
  */
 "use strict";
 
+var title_repository = 'Plone training';
+
+
+/**
+ * Return array with titles of ancestors of file.
+ * @param {number} idx - The index of the result item in global list of files
+ * @returns array
+ */
+function _getParentTitles(idx, docNames, titles) {
+  let path = docNames[idx]
+  let parentpathtokens = path.split('/').slice(0, -1);
+
+  let parentTitles = parentpathtokens.map((el, index) => {
+    let foo = `${parentpathtokens.slice(0, index+1).join('/')}`
+    let parentId = docNames.indexOf(foo);
+    if (parentId === -1) {
+      foo = `${parentpathtokens.slice(0, index+1).join('/')}/index`
+      parentId = docNames.indexOf(foo);
+    }
+    let title = parentId === -1 ? title_repository : titles[parentId];
+    return title
+  })
+
+  return parentTitles
+}
+
 /**
  * Simple result scoring code.
  */
@@ -57,6 +83,35 @@ const _removeChildren = (element) => {
 const _escapeRegExp = (string) =>
   string.replace(/[.*+\-?^${}()|[\]\\]/g, "\\$&"); // $& means the whole matched string
 
+
+function _getBreadcrumbs(item, linkUrl) {
+  // No breadcrumbs for top level pages
+  if (item[0].split('/')[1] == 'index') {
+    return null
+  }
+  let parentTitles = item[6];
+  
+  let parentDefaultTitles = [
+    "Plone trainings",
+    "Training for Plone developers"
+  ];
+  parentTitles = Array.isArray(parentTitles) ? parentTitles : parentDefaultTitles;      
+  let pathTokens = item[0].split('/')
+    .slice(0, -1);
+  let pathArray = pathTokens.map((el, index) => {
+    return {
+      "path": pathTokens.slice(0, index+1).join('/'),
+      "title": parentTitles[index]
+    }
+  })
+  let markup = pathArray
+    .map((el, idx) => {
+        return `<a href="/${el.path}">${el.title}</a>` 
+      })
+  markup.push(`<span class="lastbreadcrumb">${item[1]}</span>`)
+  return markup.join('<span class="pathseparator">&gt;</span>');
+}
+
 const _displayItem = (item, searchTerms) => {
   const docBuilder = DOCUMENTATION_OPTIONS.BUILDER;
   const docUrlRoot = DOCUMENTATION_OPTIONS.URL_ROOT;
@@ -82,6 +137,16 @@ const _displayItem = (item, searchTerms) => {
     requestUrl = docUrlRoot + docName + docFileSuffix;
     linkUrl = docName + docLinkSuffix;
   }
+
+  let breadcrumbs = _getBreadcrumbs(item, linkUrl);
+  if (breadcrumbs) {
+    let breadcrumbsNode = document.createElement("div");  
+    breadcrumbsNode.innerHTML = breadcrumbs;
+    breadcrumbsNode.classList.add("breadcrumbs");
+    listItem.appendChild(breadcrumbsNode);
+  }
+
+  // Title links to chapter
   let linkEl = listItem.appendChild(document.createElement("a"));
   linkEl.href = linkUrl + anchor;
   linkEl.dataset.score = score;
@@ -100,6 +165,7 @@ const _displayItem = (item, searchTerms) => {
       });
   Search.output.appendChild(listItem);
 };
+
 const _finishSearch = (resultCount) => {
   Search.stopPulse();
   Search.title.innerText = _("Search Results");
@@ -108,10 +174,12 @@ const _finishSearch = (resultCount) => {
       "Your search did not match any documents. Please make sure that all words are spelled correctly and that you've selected enough categories."
     );
   else
-    Search.status.innerText = _(
-      `Search finished, found ${resultCount} page(s) matching the search query.`
-    );
+    // Search.status.innerText = _(
+    //   `Search finished, found ${resultCount} page(s) matching the search query.`
+    // );
+    Search.status.innerText = `${resultCount} page(s) found.`
 };
+
 const _displayNextItem = (
   results,
   resultCount,
@@ -277,14 +345,11 @@ const Search = {
       localStorage.setItem("sphinx_highlight_terms", [...highlightTerms].join(" "))
     }
 
-    // console.debug("SEARCH: searching for:");
-    // console.info("required: ", [...searchTerms]);
-    // console.info("excluded: ", [...excludedTerms]);
-
     // array of [docname, title, anchor, descr, score, filename]
     let results = [];
     _removeChildren(document.getElementById("search-progress"));
 
+    // query matches title
     const queryLower = query.toLowerCase();
     for (const [title, foundTitles] of Object.entries(allTitles)) {
       if (title.toLowerCase().includes(queryLower) && (queryLower.length >= title.length/2)) {
@@ -297,6 +362,7 @@ const Search = {
             null,
             score,
             filenames[file],
+            _getParentTitles(file, docNames, titles),
           ]);
         }
       }
@@ -337,11 +403,7 @@ const Search = {
         return condition
       })
     }
-
-    // Enrich item with parent training title
-    // for (i = 0; i < results.length; i++)
-    //   results[i][6] = results[i][6] || 'TODO training title';
-
+    
     // now sort the results by score (in opposite order of appearance, since the
     // display function below uses pop() to retrieve items) and then
     // alphabetically
@@ -371,10 +433,6 @@ const Search = {
     }, []);
 
     results = results.reverse();
-
-    // for debugging
-    //Search.lastresults = results.slice();  // a copy
-    // console.info("search results:", Search.lastresults);
 
     // print the results
     _displayNextItem(results, results.length, searchTerms);
@@ -465,6 +523,7 @@ const Search = {
     const scoreMap = new Map();
     const fileMap = new Map();
 
+
     // perform the search on the required terms
     searchTerms.forEach((word) => {
       const files = [];
@@ -540,6 +599,7 @@ const Search = {
 
       // select one (max) score for the file.
       const score = Math.max(...wordList.map((w) => scoreMap.get(file)[w]));
+      
       // add result to the result list
       results.push([
         docNames[file],
@@ -548,6 +608,7 @@ const Search = {
         null,
         score,
         filenames[file],
+        _getParentTitles(file, docNames, titles)
       ]);
     }
     return results;