+++ CHANGES.rst
@@ -3,6 +3,12 @@ Changelog
 
 This changelog is only very rough. For the full changelog please refer to https://github.com/plone/training/commits/master
 
+
+- Enhance orientation: Add chapter title and training title in sticky top bar. @ksuess
+- Enhance orientation: Add breadcrumbs to search results. @ksuess
+
+
+
 1.2.5 (unreleased)
 ------------------